added logging to the application

+++ src/controller/task-controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { Task } from "../model/task-model";
+import logger from "../logger/logger";
 import { sequelize } from "../model/postgress.model";
 
 export const createTask = async (req, res) => {
@@ -7,6 +8,7 @@ export const createTask = async (req, res) => {
     const { title, description } = req.body;
 
     if (!title || !description) {
+      logger.error("Title or description missing when creating task");
       return res.status(400).json({
         error: true,
         statusCode: 400,
@@ -17,14 +19,16 @@ export const createTask = async (req, res) => {
     const newTask = new Task({ title, description, completed: false });
     await newTask.save();
 
-    //  const newTaskSequelize = await sequelize.models.Task.create({
-    //    title,
-    //    description,
-    //    completed: false,
-    //  });
+    const newTaskSequelize = await sequelize.models.Task.create({
+      title,
+      description,
+      completed: false,
+    });
 
+    logger.info("New task created successfully");
     res.status(200).redirect("/");
   } catch (error) {
+    logger.error("Error creating task", { error });
     return res.status(error?.status || error?.statusCode || 500).json({
       error: true,
       statusCode: error?.status || error?.statusCode || 500,
@@ -36,6 +40,7 @@ export const createTask = async (req, res) => {
 };
 
 export const renderCreateTaskPage = (req, res) => {
+  logger.info("Rendering create task page");
   res.render("create");
 };
 
@@ -43,10 +48,14 @@ export const renderEditPage = async (req, res) => {
   try {
     const task = await Task.findById(req.params.id);
     if (!task) {
+     logger.warn("Task not found while rendering edit page");
       return res.status(404).send("Task not found");
     }
+
+     logger.info("Rendering edit page for task");
     res.render("edit", { task });
   } catch (error) {
+    logger.error("error rendering edit page", { error });
     return res.status(error?.status || error?.statusCode || 500).json({
       error: true,
       statusCode: error?.status || error?.statusCode || 500,
@@ -64,9 +73,10 @@ export const editTask = async (req: Request, res: Response) => {
 
     const task = await Task.findById(taskId);
 
-    if (!task) {
-      return res.status(404).send("Task not found");
-    }
+     if (!task) {
+       logger.warn("Task not found while editing task");
+       return res.status(404).send("Task not found");
+     }
 
     task.title = title;
     task.description = description;
@@ -74,21 +84,24 @@ export const editTask = async (req: Request, res: Response) => {
 
     await task.save();
 
-    // const taskSequelize = await sequelize.models.Task.findOne({
-    //   where: { title: title }, 
-    // });
-    // if (!taskSequelize) {
-    //   return res.status(404).send("Task not found");
-    // }
+    const taskSequelize = await sequelize.models.Task.findOne({
+      where: { title: title },
+    });
+    if (!taskSequelize) {
+      logger.error("Task not found");
+      return res.status(404).send("Task not found");
+    }
 
-    //  taskSequelize.dataValues.title = title;
-    //  taskSequelize.dataValues.description = description;
-    //  taskSequelize.dataValues.completed = completed === "true";
+    taskSequelize.dataValues.title = title;
+    taskSequelize.dataValues.description = description;
+    taskSequelize.dataValues.completed = completed === "true";
 
-    //  await taskSequelize.save();
+    await taskSequelize.save();
 
+    logger.info("Task edited successfully");
     res.redirect("/");
   } catch (error) {
+    logger.error("Error editing task", { error });
     return res.status(error?.status || error?.statusCode || 500).json({
       error: true,
       statusCode: error?.status || error?.statusCode || 500,
@@ -102,9 +115,18 @@ export const editTask = async (req: Request, res: Response) => {
 export const deleteTask = async (req: Request, res: Response) => {
   try {
     const taskId = req.params.id;
+
+    if (!taskId) {
+      logger.warn("Task id  required ");
+      return res.status(404).send("Task id  required");
+    }
+
     await Task.findByIdAndDelete(taskId);
+
+     logger.info("Task deleted successfully");
     res.redirect("/");
   } catch (error) {
+    logger.error("Error deleting task", { error });
     return res.status(error?.status || error?.statusCode || 500).json({
       error: true,
       statusCode: error?.status || error?.statusCode || 500,
@@ -118,8 +140,11 @@ export const deleteTask = async (req: Request, res: Response) => {
 export const taskList = async (req: Request, res: Response) => {
   try {
     const tasks = await Task.find({}).sort({ createdAt: -1 });
+
+    logger.info("Rendering task list page");
     res.render("index", { tasks });
   } catch (error) {
+     logger.error("Error rendering task list page", { error });
     return res.status(error?.status || error?.statusCode || 500).json({
       error: true,
       statusCode: error?.status || error?.statusCode || 500,
@@ -133,9 +158,23 @@ export const taskList = async (req: Request, res: Response) => {
 export const getTask = async (req: Request, res: Response) => {
   try {
     const taskId = req.params.id;
+
+    if (!taskId) {
+      logger.warn("Taskid  required ");
+      return res.status(404).send("Task not found");
+    }
+
     const task = await Task.findById(taskId);
+
+    if (!task) {
+      logger.warn("Task not found");
+      return res.status(404).send("Task not found");
+    }
+
+     logger.info("Rendering task details page");
     res.render("task", { task });
   } catch (error) {
+    logger.error("Error rendering task details page", { error });
     return res.status(error?.status || error?.statusCode || 500).json({
       error: true,
       statusCode: error?.status || error?.statusCode || 500,
